Refactor import/export modal components and states
- Updated ExportModal and ImportModal to use new ImportExportCompletedState and ImportExportProgressState components.
- Changed import/export state imports to reflect new file structure.
- Introduced new utility functions for generating import/export messages.
- Enhanced ImportExportContentResponseBase interface to include completed and failed content details.
- Created new components: ExportInitialState, ImportExportCompletedState, and ImportExportProgressState for better modularity.
- Improved error handling in CourseModel to skip empty attachment IDs.
